rename function-utils to partial
