+++ projects/ngx-d3timeline/src/lib/core/partial.ts
